requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
audioread==3.0.0
    # via pyacoustid
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
idna==3.4
    # via requests
mutagen==1.46.0
    # via -r requirements.in
pyacoustid==1.2.2
    # via -r requirements.in
requests==2.31.0
    # via pyacoustid
urllib3==2.0.3
    # via requests
